txconfirm: fold fanout controller into an FSM the actor drives
The FeeBumpInputController turned out to be a pass-through: 21 thin
methods and three fields whose only caller was the txconfirm actor,
wrapped around an FSM that did no real work -- two lifecycle states
whose emitted outbox was discarded.

This deletes the controller and moves the whole fee-input fanout into
the FSM. The FSM's environment holds the broadcaster, so transitions do
their own wallet funding, broadcast, and reservation IO directly. That
keeps the machine at two states (idle, fanout-pending) instead of the
build/await ping-pong an outbox-per-step design would need. The only
effects expressed as outbox events are the ones that must touch
actor-owned resources: registering and tearing down the confirmation
watch, and retrying the parents that were stuck waiting on supply. The
actor drives the FSM with three events (demands-observed,
fanout-confirmed, parent-evicted) and applies the outbox it hands back.

Because protofsm tears a machine down on any transition error, an
operational fanout failure (a rejected broadcast, a rewritten output) is
stashed on the environment and surfaced to the actor after the turn
rather than returned, so the long-lived FSM survives a transient failure
and stays ready for the next demand. Behavior is unchanged; all
txconfirm tests still pass.
