+++ txconfirm/actor.go
@@ -153,12 +153,19 @@ type TxBroadcasterActor struct {
 	// submission.
 	broadcaster *CPFPBroadcaster
 
-	// feeBumpController owns the fee-input fanout lifecycle: when a CPFP
-	// child cannot find a confirmed wallet fee input, it broadcasts a
-	// fanout transaction that mints right-sized fee inputs and promotes
-	// them once it confirms. It shares the broadcaster's per-parent
-	// reservation map.
-	feeBumpController *FeeBumpInputController
+	// feeBumpFSM owns the fee-input fanout lifecycle: when a CPFP child
+	// cannot find a confirmed wallet fee input, it broadcasts a fanout
+	// transaction that mints right-sized fee inputs and promotes them once
+	// it confirms. The FSM does its own wallet/broadcast IO inside
+	// transitions and shares the broadcaster's per-parent reservation map;
+	// the actor drives it with events and applies the conf-watch / retry
+	// effects it emits as outbox events.
+	feeBumpFSM *feeBumpStateMachine
+
+	// feeBumpEnv is the fanout FSM's environment, retained so the actor can
+	// read back the per-turn operational error a transition stashed (rather
+	// than returned, which would tear the long-lived FSM down).
+	feeBumpEnv *feeBumpEnvironment
 
 	// tracked maps txid to its shared confirmation state.
 	tracked map[chainhash.Hash]*trackedTx
@@ -272,9 +279,8 @@ func NewTxBroadcasterActor(cfg Config) *TxBroadcasterActor {
 			DefaultBroadcastFailureAlertThreshold
 	}
 
-	// The fanout controller shares the broadcaster's per-parent
-	// reservation map, so it is constructed from the broadcaster rather
-	// than alongside it.
+	// The fanout FSM shares the broadcaster's per-parent reservation map,
+	// so it is constructed from the broadcaster rather than alongside it.
 	broadcaster := NewCPFPBroadcaster(BroadcasterConfig{
 		ChainSource:                    cfg.ChainSource,
 		Wallet:                         cfg.Wallet,
@@ -284,11 +290,15 @@ func NewTxBroadcasterActor(cfg Config) *TxBroadcasterActor {
 		PreSubmitTestMempoolAccept:     cfg.PreSubmitTestMempoolAccept,
 	})
 
+	log := cfg.Log.UnwrapOr(btclog.Disabled)
+	feeBumpFSM, feeBumpEnv := newFeeBumpStateMachine(broadcaster, log)
+
 	return &TxBroadcasterActor{
 		cfg:                    cfg,
-		log:                    cfg.Log.UnwrapOr(btclog.Disabled),
+		log:                    log,
 		broadcaster:            broadcaster,
-		feeBumpController:      NewFeeBumpInputController(broadcaster),
+		feeBumpFSM:             feeBumpFSM,
+		feeBumpEnv:             feeBumpEnv,
 		tracked:                make(map[chainhash.Hash]*trackedTx),
 		terminalNotifyInflight: make(map[string]struct{}),
 	}
@@ -387,7 +397,9 @@ func (a *TxBroadcasterActor) OnStop(ctx context.Context) error {
 			// tail end of a confirmation, and Evict is a
 			// no-op when parentStates has no entry.
 			a.broadcaster.Evict(ctx, entry.data.Txid)
-			a.feeBumpController.PruneParent(ctx, entry.data.Txid)
+			a.driveFeeBump(ctx, &feeBumpParentEvicted{
+				parentTxid: entry.data.Txid,
+			})
 
 			continue
 		}
@@ -409,12 +421,15 @@ func (a *TxBroadcasterActor) OnStop(ctx context.Context) error {
 		// their configured expiry fires, blocking unrelated wallet
 		// coin selection after restart.
 		a.broadcaster.Evict(ctx, entry.data.Txid)
-		a.feeBumpController.PruneParent(ctx, entry.data.Txid)
+		a.driveFeeBump(ctx, &feeBumpParentEvicted{
+			parentTxid: entry.data.Txid,
+		})
 	}
 
-	// Tear down the fanout controller's FSM goroutine alongside the
-	// per-tracked-tx FSMs.
-	a.feeBumpController.Stop()
+	// Tear down the fanout FSM goroutine alongside the per-tracked-tx FSMs.
+	if a.feeBumpFSM != nil {
+		a.feeBumpFSM.Stop()
+	}
 
 	return firstErr
 }
@@ -691,7 +706,9 @@ func (a *TxBroadcasterActor) handleCancel(ctx context.Context,
 	// Without this, a caller who cancels before confirmation can starve
 	// subsequent broadcasts of the same UTXO for up to an hour.
 	a.broadcaster.Evict(ctx, entry.data.Txid)
-	a.feeBumpController.PruneParent(ctx, entry.data.Txid)
+	a.driveFeeBump(ctx, &feeBumpParentEvicted{
+		parentTxid: entry.data.Txid,
+	})
 
 	delete(a.tracked, entry.data.Txid)
 	resp.StoppedTracking = true
@@ -1290,7 +1307,9 @@ func (a *TxBroadcasterActor) evictTerminal(ctx context.Context,
 	// any wallet-level leases held on the parent's fee UTXOs so they
 	// become immediately available to other subsystems.
 	a.broadcaster.Evict(ctx, entry.data.Txid)
-	a.feeBumpController.PruneParent(ctx, entry.data.Txid)
+	a.driveFeeBump(ctx, &feeBumpParentEvicted{
+		parentTxid: entry.data.Txid,
+	})
 
 	delete(a.tracked, entry.data.Txid)
 }
